+++ test/beans_test/dispose_destroy_all_session_test.dart
@@ -23,9 +23,12 @@ void disposeDestroyAllSession() {
 
       DDI.instance.destroyAllSession();
 
-      expect(() => DDI.instance.get<A>(), throwsA(isA<BeanNotFound>()));
-      expect(() => DDI.instance.get<B>(), throwsA(isA<BeanNotFound>()));
-      expect(() => DDI.instance.get<C>(), throwsA(isA<BeanNotFound>()));
+      expect(
+          () => DDI.instance.get<A>(), throwsA(isA<BeanNotFoundException>()));
+      expect(
+          () => DDI.instance.get<B>(), throwsA(isA<BeanNotFoundException>()));
+      expect(
+          () => DDI.instance.get<C>(), throwsA(isA<BeanNotFoundException>()));
     });
 
     test('Register, get, dispose and destroy Session bean', () {
@@ -50,9 +53,12 @@ void disposeDestroyAllSession() {
 
       DDI.instance.destroyAllSession();
 
-      expect(() => DDI.instance.get<A>(), throwsA(isA<BeanNotFound>()));
-      expect(() => DDI.instance.get<B>(), throwsA(isA<BeanNotFound>()));
-      expect(() => DDI.instance.get<C>(), throwsA(isA<BeanNotFound>()));
+      expect(
+          () => DDI.instance.get<A>(), throwsA(isA<BeanNotFoundException>()));
+      expect(
+          () => DDI.instance.get<B>(), throwsA(isA<BeanNotFoundException>()));
+      expect(
+          () => DDI.instance.get<C>(), throwsA(isA<BeanNotFoundException>()));
     });
 
     test('Register, get, dispose and destroy Session bean', () async {
@@ -80,12 +86,12 @@ void disposeDestroyAllSession() {
 
       DDI.instance.destroyAllSession();
 
-      expect(
-          () async => DDI.instance.getAsync<A>(), throwsA(isA<BeanNotFound>()));
-      expect(
-          () async => DDI.instance.getAsync<B>(), throwsA(isA<BeanNotFound>()));
-      expect(
-          () async => DDI.instance.getAsync<C>(), throwsA(isA<BeanNotFound>()));
+      expect(() async => DDI.instance.getAsync<A>(),
+          throwsA(isA<BeanNotFoundException>()));
+      expect(() async => DDI.instance.getAsync<B>(),
+          throwsA(isA<BeanNotFoundException>()));
+      expect(() async => DDI.instance.getAsync<C>(),
+          throwsA(isA<BeanNotFoundException>()));
     });
   });
 }