+++ test/beans_test/module_dependent_test.dart
@@ -22,7 +22,8 @@ void moduleDependentTest() {
 
       DDI.instance.destroy<ModuleDependent>();
 
-      expect(() => DDI.instance.get<C>(), throwsA(isA<BeanNotFound>()));
+      expect(
+          () => DDI.instance.get<C>(), throwsA(isA<BeanNotFoundException>()));
     });
   });
 }